Create crawling_man.cpp
Source code for program seen in these youtube videos:
https://www.youtube.com/watch?v=i2VvvOWWquo
https://www.youtube.com/watch?v=ULEwgWRLihs
https://www.youtube.com/watch?v=Gciis_WOodo

To compile: download and setup OpenCV 3.4 and Dlib 18.18 (later will probably also work if you tweak the rest of the stuff.)

Make a Visual Studio project and add this C++ source code. In the project project configuration, also add the directories for Dlib to the include files, and the right OpenCV locations for the include, libraries, and executable directories. In the linker settings, it's easiest for me just to tell you to add all the opencv .lib files.

If you can compile and run the sample code, you can compile and run this.

Enjoy.